# docs/swagger.yaml
basePath: /api
definitions:
...
  apierror.ApiError:
    properties:
      code:
        type: integer
      error: {}
    type: object
...
  db.Course:
    properties:
      created_at:
        $ref: '#/definitions/pgtype.Timestamp'
      description:
        $ref: '#/definitions/pgtype.Text'
      id:
        type: integer
      join_code:
        type: string
      name:
        type: string
      teacher_id:
        type: integer
      updated_at:
        $ref: '#/definitions/pgtype.Timestamp'
    type: object
...
  db.User:
    properties:
      created_at:
        $ref: '#/definitions/pgtype.Timestamp'
      elo:
        type: integer
      email:
        type: string
      id:
        type: integer
      password_hash:
        type: string
      roles:
        type: integer
      updated_at:
        $ref: '#/definitions/pgtype.Timestamp'
      username:
        type: string
    type: object
...
  models.CourseResponse:
    properties:
      assignment_count:
        type: integer
      created_at:
        $ref: '#/definitions/pgtype.Timestamp'
      description:
        $ref: '#/definitions/pgtype.Text'
      id:
        type: integer
      join_code:
        type: string
      name:
        type: string
      student_count:
        type: integer
      teacher:
        $ref: '#/definitions/db.User'
      teacher_id:
        type: integer
      updated_at:
        $ref: '#/definitions/pgtype.Timestamp'
    type: object
...
  models.CreateCourseDTO:
    properties:
      description:
        maxLength: 1000
        type: string
      name:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - name
    type: object
...
  models.GradeSummary:
    properties:
      assignment_id:
        type: integer
      assignment_title:
        type: string
      average_grade:
        type: number
      tasks_count:
        type: integer
    type: object
  models.JoinCourseDTO:
    properties:
      join_code:
        maxLength: 10
        minLength: 6
        type: string
    required:
    - join_code
    type: object
...
  models.StudentInCourse:
    properties:
      created_at:
        $ref: '#/definitions/pgtype.Timestamp'
      elo:
        type: integer
      email:
        type: string
      enrolled_at:
        type: string
      grades:
        items:
          $ref: '#/definitions/models.GradeSummary'
        type: array
      id:
        type: integer
      password_hash:
        type: string
      role:
        type: string
      roles:
        type: integer
      updated_at:
        $ref: '#/definitions/pgtype.Timestamp'
      username:
        type: string
    type: object
...
  models.UpdateCourseDTO:
    properties:
      description:
        maxLength: 1000
        type: string
      name:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - name
    type: object
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
...
  pgtype.Text:
    properties:
      string:
        type: string
      valid:
        type: boolean
    type: object
  pgtype.Timestamp:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        description: Time zone will be ignored when encoding to PostgreSQL.
        type: string
      valid:
        type: boolean
    type: object
...
host: localhost:4000
info:
  contact: {}
  title: Rankode
  version: "1.0"
paths:
...
  /courses:
    get:
      consumes:
      - application/json
      description: Retrieves courses for the current user (as teacher or student)
      produces:
      - application/json
      responses:
        "200":
          description: List of courses
          schema:
            items:
              $ref: '#/definitions/db.Course'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierror.ApiError'
      security:
      - ApiKeyAuth: []
      summary: List courses
      tags:
      - Courses
    post:
      consumes:
      - application/json
      description: Creates a new course (teacher only)
      parameters:
      - description: Course creation payload
        in: body
        name: course
        required: true
        schema:
          $ref: '#/definitions/models.CreateCourseDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created course
          schema:
            $ref: '#/definitions/db.Course'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "403":
          description: Forbidden (not a teacher)
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierror.ApiError'
      security:
      - ApiKeyAuth: []
      summary: Create a new course
      tags:
      - Courses
  /courses/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a course (teacher only)
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Course deleted successfully
        "403":
          description: Forbidden (not teacher of course)
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "404":
          description: Course not found
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierror.ApiError'
      security:
      - ApiKeyAuth: []
      summary: Delete course
      tags:
      - Courses
    get:
      consumes:
      - application/json
      description: Retrieves a single course by its ID
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Course details
          schema:
            $ref: '#/definitions/models.CourseResponse'
        "404":
          description: Course not found
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierror.ApiError'
      security:
      - ApiKeyAuth: []
      summary: Get course by ID
      tags:
      - Courses
    put:
      consumes:
      - application/json
      description: Updates an existing course (teacher only)
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      - description: Course update payload
        in: body
        name: course
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCourseDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Course updated successfully
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "403":
          description: Forbidden (not teacher of course)
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "404":
          description: Course not found
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierror.ApiError'
      security:
      - ApiKeyAuth: []
      summary: Update course
      tags:
      - Courses
  /courses/{id}/enroll:
    delete:
      consumes:
      - application/json
      description: Removes the current user from a course
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully unenrolled
        "400":
          description: Not enrolled
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "404":
          description: Course not found
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierror.ApiError'
      security:
      - ApiKeyAuth: []
      summary: Unenroll from a course
      tags:
      - Courses
    post:
      consumes:
      - application/json
      description: Enrolls the current user in a course (requires teacher permission?)
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully enrolled
        "400":
          description: Already enrolled
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "404":
          description: Course not found
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierror.ApiError'
      security:
      - ApiKeyAuth: []
      summary: Enroll in a course (direct)
      tags:
      - Courses
  /courses/{id}/students:
    get:
      consumes:
      - application/json
      description: Lists all students enrolled in a course (teacher only)
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of students
          schema:
            items:
              $ref: '#/definitions/models.StudentInCourse'
            type: array
        "403":
          description: Forbidden (not teacher of course)
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "404":
          description: Course not found
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierror.ApiError'
      security:
      - ApiKeyAuth: []
      summary: List students in course
      tags:
      - Courses
  /courses/enroll:
    post:
      consumes:
      - application/json
      description: Enrolls the current user in a course using join code
      parameters:
      - description: Join code
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.JoinCourseDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully enrolled
          schema:
            $ref: '#/definitions/db.Course'
        "400":
          description: Already enrolled or invalid join code
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "404":
          description: Course not found
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierror.ApiError'
      security:
      - ApiKeyAuth: []
      summary: Enroll in a course using join code
      tags:
      - Courses
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"